chore: split exclusions
